requirements.txt
Django==1.6.3
Markdown==2.4.1
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.3.0
dj-static==0.0.5
django-crispy-forms==1.5.0
django-filter==0.7
django-registration==1.0
django-toolbelt==0.0.1
djangorestframework==2.3.14
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
